Added Change Location href
new href added to nav bar
on click if location can't be changed on device show error.
if it can be prompt user if they want to change current location,
if no, do nothing.
if yes, get location, set location to cookie for 7 days.
Refresh page so that new marker is reflected.